backend/requirements.txt
anthropic==0.40.0
Flask==3.1.0
flask-cors==5.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
